+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+
+## [0.1.2]
 ### Added
 - Added `html`, `colorpicker`, `datepicker` field
 
@@ -12,5 +14,6 @@
 ## [0.1.0] - 2017-09-19
 - First release
 
-[Unreleased]: https://github.com/truongwp/puppyfw/compare/0.1.1...HEAD
-[0.1.1]: https://github.com/truongwp/puppyfw/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/truongwp/puppyfw/compare/0.1.2...HEAD
+[0.1.2]: https://github.com/truongwp/puppyfw/compare/0.1.1...0.1.2
+[0.1.1]: https://github.com/truongwp/puppyfw/compare/0.1.0...0.1.1